Add a "lite" version of the library (via omitted feature) with most deps removed.
RIght now, egg includes quite a lot of useful functionality, but this
comes at the cost of a somewhat non-trivial dependency tree. (It's
certainly not hundreds-of-crates "huge", but it is still a lot.)
Sometimes, it is desirable to cut down on dependencies as much as
possible. For example, the user may want to use the core egraph data
structure but handle most everything else in a custom way, and might not
want to pull in egg's choice of utility crates for symbol interning,
parsing, error handling, timing, logging, and all the rest, especially
if they already have ways of handling these needs. This is somewhat more
likely to be the case for compiler-like use-cases: For example,
Cranelift has an explicit policy to reduce dependencies in crates
low in the dependency DAG, both for build-time and trusted-base-size
reasons.

This PR introduces a `standard` feature (on by default) that gates a
large fraction of the library. With the feature disabled, one gets the
`EGraph`, `EClass`, and `Language` types, and other associated bits, but
no rewrite rules or pattern matching, no runners, no extractors, no
explanations, and no other auxiliary features.

By default, building `egg` builds 36 crates currently, taking 11.72s on
my machine for a release build. With this change, building `egg` builds
20 crates, taking 4.11s on my machine. The only remaining dependencies
are `fxhash`, `indexmap`, `hashbrown`, and `smallvec`; in addition to
this list being quite short, these crates are often already in the
user's dependency DAG.  Furthermore, and importantly for build time,
there are no remaining proc-macros at all.

The original `cargo tree` looks like (excluding dev-dependencies):

```
egg v0.8.1
├── fxhash v0.2.1
│   └── byteorder v1.4.3
├── hashbrown v0.12.1
│   └── ahash v0.7.6
│       ├── getrandom v0.2.6
│       │   ├── cfg-if v1.0.0
│       │   └── libc v0.2.126
│       └── once_cell v1.12.0
│       [build-dependencies]
│       └── version_check v0.9.4
├── indexmap v1.8.1
│   └── hashbrown v0.11.2
│   [build-dependencies]
│   └── autocfg v1.1.0
├── instant v0.1.12
│   └── cfg-if v1.0.0
├── log v0.4.17
│   └── cfg-if v1.0.0
├── smallvec v1.8.0
├── symbol_table v0.1.0
│   └── hashbrown v0.12.1 (*)
├── symbolic_expressions v5.0.3
└── thiserror v1.0.31
    └── thiserror-impl v1.0.31 (proc-macro)
        ├── proc-macro2 v1.0.39
        │   └── unicode-ident v1.0.0
        ├── quote v1.0.18
        │   └── proc-macro2 v1.0.39 (*)
        └── syn v1.0.95
            ├── proc-macro2 v1.0.39 (*)
            ├── quote v1.0.18 (*)
            └── unicode-ident v1.0.0
```

with this PR and `--no-default-features`, it becomes:

```
egg v0.8.1
├── fxhash v0.2.1
│   └── byteorder v1.4.3
├── hashbrown v0.12.1
│   └── ahash v0.7.6
│       ├── getrandom v0.2.6
│       │   ├── cfg-if v1.0.0
│       │   └── libc v0.2.126
│       └── once_cell v1.12.0
│       [build-dependencies]
│       └── version_check v0.9.4
├── indexmap v1.8.1
│   └── hashbrown v0.11.2
│   [build-dependencies]
│   └── autocfg v1.1.0
└── smallvec v1.8.0
```